requirements.txt
flask
openai
pytesseract
Pillow